NOTFORUPSREAM: x86/sgx: Temporary user space policy stopgap
Enclave pages are added to an initialized enclave with default RW
permissions and not allowed to exceed these permissions during their
lifetime.

There are some use cases that require dynamically added pages to obtain
executable permissions and that is not possible with the current
implementation.

Supporting dynamically added pages to obtain executable permission require
integration with user space policy that does not yet exist.

As a temporary stopgap, allow dynamically allowed pages to obtain
executable permissions in support of these use cases.

This is not the solution that will be targeted for upstream inclusion.